RGBImageRenderer/pyproject.toml
[tool.poetry]
name = "RGBRenderer"
version = "0.1.0"
description = "Render images using configurable effects"
authors = [
  "Nicholas Lee <csudcy@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.7,<3.9"  # Compatible python versions must be declared here
click = "^7.1.2"
Pillow = "^8.0.1"

[tool.poetry.dev-dependencies]
isort = "^5.6.4"
pytype = "^2020.10.8"
yapf = "^0.30.0"
